Complete turret authorization workflow (#14)
* Implement turret authorization Phase 5

* feat: complete turret authorization workflow

* chore: tidy up

* Implement deploy workflow tabs

* Update E2E workflow selectors

* fix: headers

* Persist settled compilation output

* Auto-detect Vault world target

* Fix React Flow update loop

* Reduce Move diagnostic noise in footer

* Add visual deployment target selector

* Improve deployment flow and world contract compilation

* chore: tidy up

* fix: resolve all ESLint warnings across 6 files

- App.tsx: extract compilation state handler to reduce StandardApp line count
- emitter.ts: extract BCS unpacker into emitBcsUnpackerFunctions helper
- moveCompiler.ts: extract build result handlers to reduce compileMove complexity
- AuthorizeView.tsx: wrap selectedTurretIds in useMemo for stable hook deps
- VisualDeploymentTargetSelector.tsx: extract TargetMenu sub-component
- compilationStateStorage.ts: extract type guard helpers and optional field
  parsers to reduce cyclomatic complexity in parseArtifact, parseDiagnostics,
  and parseCompilationStatus

* Refine deployment flows and document ADR 009

* Remove redundant Move source header chip

* Fix deployment preview regression paths

* fix: address PR #14 review comments

+++ src/compiler/generators/dataAccessors.ts
@@ -73,8 +73,8 @@ function createGetTribeAccessor(): NodeCodeGenerator {
 
       return [
         ...createCommentBlock(node, ["accessor getTribe", "extract tribe and owner tribe"]),
-        { code: `let ${tribeBinding}: u64 = ${targetBinding}.character_tribe;`, nodeId: node.id, indent: 2 },
-        { code: `let ${ownerTribeBinding}: u64 = owner_tribe;`, nodeId: node.id, indent: 2 },
+        { code: `let ${tribeBinding}: u32 = ${targetBinding}.character_tribe;`, nodeId: node.id, indent: 2 },
+        { code: `let ${ownerTribeBinding}: u32 = owner_tribe;`, nodeId: node.id, indent: 2 },
       ];
     },
   };
@@ -152,7 +152,7 @@ const dataAccessorGenerators: readonly NodeCodeGenerator[] = [
 
       return [
         ...createCommentBlock(node, ["accessor getBehaviour", "read behaviour code"]),
-        { code: `let ${outputBinding}: u64 = ${targetBinding}.behaviour_change;`, nodeId: node.id, indent: 2 },
+        { code: `let ${outputBinding}: u8 = ${targetBinding}.behaviour_change;`, nodeId: node.id, indent: 2 },
       ];
     },
   },
